idr0011: *manually* add fields
Five fields were not picked up by the make_screen.py script
because they lacked a well name. These were provided via email
by the author:

```
    1 Stinger plates-extra images screen T34 x TS-more repeats/Plate1-Green-B-TS-Stinger/Plate1-Green-B-27-Scene-2-P1-02.czi

This one is B6 so Plate1-Green-B-B6-02

      1 Stinger plates-extra images screen T34 x TS-more repeats/Plate1-Green-B-TS-Stinger/Plate1-Green-B-27-Scene-3-P2-03.czi

This one is B6 aswell so Plate1-Green-B-B6-03

      1 Stinger plates-extra images screen T34 x TS-more repeats/Plate1-Green-B-TS-Stinger/Plate1-Green-B-36-Scene-3-P2-03.czi

This one is F8 so Plate1-Green-B-F8-03

      1 Stinger plates-extra images screen T34 x TS-more repeats/Plate1-Red-B-TS-Stinger/Plate1-Red-B-18-Scene-3-P3-03.czi

This one is B4 so Plate1-Red-B-B4-03

      1 Stinger plates-extra images screen T34 x TS-more repeats/Plate1-Red-B-TS-Stinger/Plate1-Red-B-24-Scene-3-P3-03.czi

This one is E6 so Plate1-Red-B-E6-03

These would have not been included in the analysis because the script used the
P number as position and we excluded them but these are the strains these
images correspond to.

```