idr0010: fix weird channel names
Each of these plates has a discrepancy between the filenames for
each of the channels (extra dash, wrong integer). Previously,
work by Simone was intended to allow detecting these cases, but
at the moment this is not functioning. Since all the of channel
names will need to be reset, further modifying them in this PR
seems the lesser evil.

See https://trello.com/c/vMCDDpwY/11-dna-damage-import for more info.