Create new symlink for idr0036 for file path preservation on live resource
The raw imaging data import of idr0036 was performed with a folder suffixed
with compoundprofiling. This suffix was later changed to cellpainting. While
the raw TIFF files on disk are all under a cellpainting fileset, putting this
symlink under version control is mandatory so that a clean checkout of
idr-metadata does not break the existing pixel data.